+++ js/main.js
@@ -1,6 +1,3 @@
-import './setup.js';
 import {getPictures} from './data.js';
 
-console.log(
-  getPictures()
-);
+getPictures();
