+++ js/util.js
@@ -12,4 +12,15 @@ const getRandomInteger = (a, b) => {
 const getRandomArrayElement = (array) =>
   array[getRandomInteger(0, array.length - 1)];
 
-export {getRandomInteger, getRandomArrayElement};
+// Вспомогательная функция #3 - генерация ID
+
+function createIdGenerator () {
+  let lastGeneratedId = 0;
+
+  return () => {
+    lastGeneratedId += 1;
+    return lastGeneratedId;
+  };
+}
+
+export {getRandomInteger, getRandomArrayElement, createIdGenerator};
